requirements.txt
annotated-types==0.7.0
cachetools==5.5.0
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
coverage==7.6.1
cssselect2==0.7.0
defusedxml==0.7.1
Faker==29.0.0
google-api-core==2.20.0
google-auth==2.35.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-secret-manager==2.20.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpc-google-iam-v1==0.13.1
grpcio==1.66.2
grpcio-status==1.66.2
idna==3.10
iniconfig==2.0.0
linkify-it-py==2.0.3
markdown-it-py==3.0.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
mslex==1.2.0
packaging==24.1
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
proto-plus==1.24.0
protobuf==5.28.2
psutil==5.9.8
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pyfiglet==1.0.2
Pygments==2.18.0
pytest==8.3.3
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
rich==13.8.1
rsa==4.9
ruff==0.6.4
six==1.16.0
taskipy==1.13.0
textual==0.81.0
tinycss2==1.3.0
tomli==2.0.1
typing_extensions==4.12.2
uc-micro-py==1.0.3
urllib3==2.2.3
webencodings==0.5.1